pyproject.toml
[project]
name = "agentic-trading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alpaca-py>=0.39.1",
    "alpaca-trade-api>=3.2.0",
    "dotenv>=0.9.9",
    "ipython>=9.1.0",
    "langchain>=0.3.23",
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.3.21",
    "langchain-huggingface>=0.1.2",
    "langchain-ollama>=0.3.1",
    "langgraph>=0.3.28",
    "langgraph-cli[inmem]>=0.2.3",
    "pypdf>=5.4.0",
    "tiktoken>=0.9.0",
]
